Ajustes e implementaçao do reset de senha.
