+++ src/main/java/br/com/api/auth/dto/ResetPasswordDto.java
@@ -6,7 +6,9 @@ import lombok.Data;
 public class ResetPasswordDto {
 
     private Long id;
-    private String senha;
+    private String password;
     private String email;
-    private String nome;
+    private String userName;
+    private String cpf;
+
 }
