+++ src/main/resources/application.properties
@@ -67,7 +67,7 @@ email.broker.queue.email.name=default.email
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
 spring.mail.username=arlindoldev@gmail.com
-spring.mail.password= wlnz oavo **** wbzc
+spring.mail.password= wlnz oavo nhuu wbzc
 
 spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
